refactor: Create SearchBar component
